requirements.txt
streamlit
python-dotenv
requests
faiss-cpu
google-generativeai
langchain
langchain-google-genai
langchain-community
pymupdf